+++ src/App.vue
@@ -17,7 +17,7 @@
 
     <modal-loader />
     <alert-snackbar />
-    <update-check v-if="buildConfig.tauri" />
+    <tauri-update-check v-if="buildConfig.tauri" />
 
     <app-footer />
   </q-layout>
@@ -34,7 +34,7 @@
   import { useThemeStore } from './store/theme.ts'
   import { useConnectionStore } from './store/connection'
   import { setAppThemeCss, setupThemeListener } from './helpers/theme.ts'
-  import UpdateCheck from './components/base/UpdateCheck.vue'
+  import TauriUpdateCheck from './components/base/TauriUpdateCheck.vue'
   import { buildConfig } from './buildConfig.ts'
 
   const themeStore = useThemeStore()